@@ -5,7 +5,7 @@ homepage: https://github.com/rob333/styled
 
 environment:
   sdk: '>=3.1.0 <4.0.0'
-  flutter: ">=3.0.0"
+  flutter: ">=3.13.0"
 
 dependencies:
   flutter:
