@@ -18,17 +18,24 @@ customMultiChildLayoutd(List argsList, Map<Symbol, dynamic> origArgsMap) {
     }
   }
 
-  argsMap[#children] = widgets;
-
-  // named args(origArgsMap) precedes positional ones
+  // named args(origArgsMap) precede positional ones
   if (origArgsMap.isNotEmpty) {
     argsMap.addAll(origArgsMap);
+
+    // merge widgets
+    final list = argsMap[#children] as List<Widget>?;
+    if (list != null) {
+      widgets.addAll(list);
+    }
   }
+
+  argsMap[#children] = widgets;
 
   //* required:
   assert(() {
     if (argsMap[#delegate] == null) {
-      throw FlutterError("CustomMultiChildLayoutd requires `delegate:MultiChildLayoutDelegate`");
+      throw FlutterError(
+          "CustomMultiChildLayoutd requires `delegate:MultiChildLayoutDelegate`");
     }
     return true;
   }());
