@@ -55,17 +55,24 @@ gridViewd(List argsList, Map<Symbol, dynamic> origArgsMap) {
     }
   }
 
-  argsMap[#children] = widgets;
-
   // named args(origArgsMap) precedes positional ones
   if (origArgsMap.isNotEmpty) {
     argsMap.addAll(origArgsMap);
+
+    // merge widgets
+    final list = argsMap[#children] as List<Widget>?;
+    if (list != null) {
+      widgets.addAll(list);
+    }
   }
+
+  argsMap[#children] = widgets;
 
   //* required:
   assert(() {
     if (argsMap[#gridDelegate] == null) {
-      throw FlutterError("GridViewd requires `gridDelegate:SliverGridDelegate`");
+      throw FlutterError(
+          "GridViewd requires `gridDelegate:SliverGridDelegate`");
     }
     return true;
   }());
