@@ -25,13 +25,18 @@ indexedStackd(List argsList, Map<Symbol, dynamic> origArgsMap) {
         argsMap[#index] = arg;
     }
   }
-
-  argsMap[#children] = widgets;
-
-  // named args(origArgsMap) precedes positional ones
+  // named args(origArgsMap) precede positional ones
   if (origArgsMap.isNotEmpty) {
     argsMap.addAll(origArgsMap);
+
+    // merge widgets
+    final list = argsMap[#children] as List<Widget>?;
+    if (list != null) {
+      widgets.addAll(list);
+    }
   }
+
+  argsMap[#children] = widgets;
 
   return Function.apply(IndexedStack.new, [], argsMap);
 }
